type chnges

+++ pojo-ops/src/main/java/com/stanzaliving/core/cafe/dto/CafeVasItemDto.java
@@ -22,7 +22,7 @@ public class CafeVasItemDto {
     private boolean containsEgg;
     private boolean nonVeg;
     private boolean specialVeg;
-    private Integer shelfLife;
+    private Double shelfLife;
     private String description;
     private String image;
     private Integer serves;
